Update mantid dependency to latest Mantid stable 6.11 (#15)
* Update Mantid and MantidWorkbench versions to 6.11

Upgraded Mantid in conda.recipe meta.yaml from 6.8.0 to 6.11 and MantidWorkbench in environment.yml from 6.9.1 to 6.11. This ensures compatibility with the latest features and fixes provided in version 6.11.

Signed-off-by: Jose Borreguero <borreguero@gmail.com>

* Update release notes with latest PR changes

This commit updates the release notes to include the latest changes from PR 15. It also corrects the order of previous PRs for better chronological tracking.

Signed-off-by: Jose Borreguero <borreguero@gmail.com>

---------

Signed-off-by: Jose Borreguero <borreguero@gmail.com>